@@ -101,7 +101,9 @@ export default function firebaseConnect(dataOrFn = []) {
         unWatchEvents(firebase, dispatch, this._firebaseEvents)
       }
 
-      componentWillReceiveProps(np) {
+      /* eslint-disable camelcase */
+      UNSAFE_componentWillReceiveProps(np) {
+        /* eslint-enable camelcase */
         const { firebase, dispatch } = this.props
         const inputAsFunc = createCallable(dataOrFn)
         const data = inputAsFunc(np, this.store)
