@@ -3,17 +3,12 @@ import {
   set,
   get,
   has,
-  last,
   map,
   mapValues,
-  every,
   reduce,
   defaultsDeep,
-  isString,
   compact,
-  some,
-  isArray,
-  isFunction
+  some
 } from 'lodash'
 import { topLevelPaths } from './constants'
 import { getPopulateObjs } from './utils/populate'
@@ -126,7 +121,7 @@ export function getVal(firebase, path, notSetValue) {
 export function isLoaded(...args) {
   return !args || !args.length
     ? true
-    : every(args, arg => arg !== undefined && get(arg, 'isLoaded') !== false)
+    : args.every(arg => arg !== undefined && get(arg, 'isLoaded') !== false)
 }
 
 /**
@@ -169,7 +164,9 @@ export function isLoaded(...args) {
  * export default enhance(Todos)
  */
 export function isEmpty(...args) {
-  return some(args, arg => !(arg && size(arg)) || arg.isEmpty === true)
+  return !args || !args.length
+    ? true
+    : args.some(arg => !(arg && size(arg)) || arg.isEmpty === true)
 }
 
 /**
@@ -189,15 +186,15 @@ export function fixPath(path) {
  * @param {Object} list - Path of parameter to load
  * @param {Object} populateSettings - Object with population settings
  */
-const buildChildList = (state, list, p) =>
-  mapValues(list, (val, key) => {
+function buildChildList(state, list, p) {
+  return mapValues(list, (val, key) => {
     let getKey = val
     // Handle key: true lists
     if (val === true || p.populateByKey) {
       getKey = key
     }
     // Allow for aliasing populated data see #126 for more details
-    const dotRoot = compact(p.root.split('/')).join('.')
+    const dotRoot = getDotStrPath(p.root)
     const pathArr = [dotRoot, getKey]
 
     // Handle child param
@@ -216,10 +213,11 @@ const buildChildList = (state, list, p) =>
     // Populate child does not exist
     return val === true || p.populateByKey ? val : getKey
   })
+}
 
 /**
  * @private
- * @description Populate a child based on config. Handles list population
+ * Populate a child based on config. Handles list population
  * by making use of buildChildList.
  * @param {Object} state - Firebase state object
  * @param {Object} child - Path of parameter to load
@@ -232,9 +230,12 @@ function populateChild(state, child, p) {
     return null
   }
   // populate child is key
-  if (isString(childVal)) {
+  if (typeof childVal === 'string' || childVal instanceof String) {
     // attach child paramter if it exists
-    const dotRoot = compact(p.root.split('/')).join('.')
+    const dotRoot = p.root
+      .split('/')
+      .filter(Boolean) // Drop falsey values (compact)
+      .join('.')
     const pathArr = [dotRoot, childVal]
 
     // Handle child param
@@ -260,12 +261,12 @@ function populateChild(state, child, p) {
 }
 
 /**
- * @description Populate with data from redux.
+ * Populate with data from multiple locations of redux state.
  * @param {Object} state - Firebase state object (state.firebase in redux store)
  * @param {String} path - Path of parameter to load
  * @param {Array} populates - Array of populate config objects
  * @param {Object|String|Boolean} notSetValue - Value to return if value is not found
- * @return {Object} Data located at path within Immutable Object
+ * @returns {Object} Data located at path within Immutable Object
  * @example <caption>Basic</caption>
  * import { compose } from 'redux'
  * import { connect } from 'react-redux'
@@ -308,10 +309,12 @@ export function populate(state, path, populates, notSetValue) {
 
   // check for if data is single object or a list of objects
   const populatesForData = getPopulateObjs(
-    isFunction(populates) ? populates(last(pathArr), data) : populates
+    typeof populates === 'function'
+      ? populates(pathArr.slice(-1)[0], data) // pass last slice in path
+      : populates
   )
 
-  if (isArray(data)) {
+  if (Array.isArray(data)) {
     // When using a path in ordered, data will be an array instead of an object
     // and data is located at the `value` prop
     const someArrayItemHasKey = array => key =>
@@ -362,7 +365,7 @@ export function populate(state, path, populates, notSetValue) {
     const key = pathArr[0] === 'ordered' ? child.key : childKey
     // get populate settings on item level (passes child if populates is a function)
     const populatesForDataItem = getPopulateObjs(
-      isFunction(populates) ? populates(key, child) : populates
+      typeof populates === 'function' ? populates(key, child) : populates
     )
     // confirm at least one populate value exists on child
     const dataHasPopulateChilds = some(populatesForDataItem, p =>
