@@ -1,4 +1,4 @@
-import createFirebaseInstance from './createFirebaseInstance'
+import createFirebaseInstance, { getFirebase } from './createFirebaseInstance'
 import ReactReduxFirebaseProvider from './ReactReduxFirebaseProvider'
 import ReactReduxFirebaseContext from './ReactReduxFirebaseContext'
 import ReduxFirestoreProvider from './ReduxFirestoreProvider'
@@ -37,7 +37,7 @@ export default {
   firebaseStateReducer: reducer,
   constants,
   actionTypes,
+  getFirebase,
   authIsReady,
-  helpers,
   ...helpers
 }
