@@ -1,4 +1,4 @@
-import { flatMap, isArray, isObject, isString, remove } from 'lodash'
+import { flatMap, isObject, remove } from 'lodash'
 import { getPopulates } from './populate'
 import { getQueryIdFromPath } from './query'
 
@@ -38,11 +38,12 @@ export function pathStrToObj(path) {
  */
 export function getEventsFromInput(paths) {
   return flatMap(paths, path => {
-    if (isString(path)) {
+    // If path is a string - convert to obj and place within new array
+    if (typeof path === 'string' || path instanceof String) {
       return [pathStrToObj(path)]
     }
 
-    if (isArray(path)) {
+    if (Array.isArray(path)) {
       // TODO: Handle input other than array with string
       // TODO: Handle populates within array
       return [
